provide runner to workers, use it for all builds
